business_llm_assistant/requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
llama-cpp-python>=0.1.77
langchain>=0.0.200
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Database and Storage
SQLAlchemy>=2.0.0
pymongo>=4.4.0
faiss-gpu>=1.7.4; platform_system != "Windows"
redis>=4.5.0
diskcache>=5.4.0

# Document Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web and API
Flask>=2.2.0
Flask-RESTful>=0.3.9
pydantic>=2.0.0
requests>=2.28.0
streamlit>=1.25.0; platform_system != "Windows"
fastapi>=0.100.0
uvicorn>=0.23.0

# Voice Interface (Optional)
pyttsx3>=2.90
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
whisper>=1.0

# System Utilities
psutil>=5.9.0
tqdm>=4.65.0
schedule>=1.2.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Security
cryptography>=40.0.0
bcrypt>=4.0.0
python-jose>=3.3.0
passlib>=1.7.4

# MCP Integration
websockets>=11.0.0
aiohttp>=3.8.4

# Optional Dependencies
# Uncomment as needed
# tensorboard>=2.12.0  # For training visualization
# gradio>=3.35.0  # For creating simple web interfaces
# chromadb>=0.4.0  # Alternative vector store
# pinecone-client>=2.2.0  # Cloud vector database
# anthropic>=0.3.0  # For Claude API support
